+++ Pal/include/arch/x86_64/cpu.h
@@ -31,6 +31,40 @@ static inline uint64_t get_tsc(void) {
     return lo | ((uint64_t)hi << 32);
 }
 
+/*!
+ * \brief Low-level wrapper around RDRAND instruction (get hardware-generated random value).
+ */
+static inline uint32_t rdrand(void) {
+    uint32_t ret;
+    __asm__ volatile(
+        "1: .byte 0x0f, 0xc7, 0xf0\n" /* RDRAND %EAX */
+        "jnc 1b\n"
+        :"=a"(ret)
+        :: "cc");
+    return ret;
+}
+
+/*!
+ * \brief Low-level wrapper around RDFSBASE instruction (read FS register; allowed in enclaves).
+ */
+static inline uint64_t rdfsbase(void) {
+    uint64_t fsbase;
+    __asm__ volatile(
+        ".byte 0xf3, 0x48, 0x0f, 0xae, 0xc0\n" /* RDFSBASE %RAX */
+        : "=a"(fsbase) :: "memory");
+    return fsbase;
+}
+
+/*!
+ * \brief Low-level wrapper around WRFSBASE instruction (modify FS register; allowed in enclaves).
+ */
+static inline void wrfsbase(uint64_t addr) {
+    __asm__ volatile(
+        ".byte 0xf3, 0x48, 0x0f, 0xae, 0xd7\n" /* WRFSBASE %RDI */
+        :: "D"(addr) : "memory");
+}
+
+
 #define CPU_RELAX() __asm__ volatile("pause")
 
 /* some non-Intel clones support out of order store; WMB() ceases to be a nop for these */
