+++ Pal/src/host/Linux-SGX/db_main-x86_64.c
@@ -6,6 +6,7 @@
  */
 
 #include "api.h"
+#include "cpu.h"
 #include "linux_utils.h"
 #include "pal.h"
 #include "pal_debug.h"
