+++ Pal/src/host/Linux-SGX/sgx_api.h
@@ -51,37 +51,4 @@ static inline int64_t sgx_getkey(sgx_key_request_t* keyrequest, sgx_key_128bit_t
     return rax;
 }
 
-/*!
- * \brief Low-level wrapper around RDRAND instruction (get hardware-generated random value).
- */
-static inline uint32_t rdrand(void) {
-    uint32_t ret;
-    __asm__ volatile(
-        "1: .byte 0x0f, 0xc7, 0xf0\n" /* RDRAND %EAX */
-        "jnc 1b\n"
-        :"=a"(ret)
-        :: "cc");
-    return ret;
-}
-
-/*!
- * \brief Low-level wrapper around RDFSBASE instruction (read FS register; allowed in enclaves).
- */
-static inline uint64_t rdfsbase(void) {
-    uint64_t fsbase;
-    __asm__ volatile(
-        ".byte 0xf3, 0x48, 0x0f, 0xae, 0xc0\n" /* RDFSBASE %RAX */
-        : "=a"(fsbase));
-    return fsbase;
-}
-
-/*!
- * \brief Low-level wrapper around WRFSBASE instruction (modify FS register; allowed in enclaves).
- */
-static inline void wrfsbase(uint64_t addr) {
-    __asm__ volatile(
-        ".byte 0xf3, 0x48, 0x0f, 0xae, 0xd7\n" /* WRFSBASE %RDI */
-        :: "D"(addr));
-}
-
 #endif /* SGX_API_H */
