+++ Pal/src/host/Linux/db_pipes.c
@@ -7,6 +7,7 @@
 
 #include <asm/errno.h>
 #include <asm/fcntl.h>
+#include <asm/ioctls.h>
 #include <asm/poll.h>
 #include <linux/time.h>
 #include <linux/types.h>
